validate that figures, tables, formulae, annexes are cross-referenced: https://github.com/metanorma/metanorma-iso/issues/1211
